Separating factory from transformer and using existing content handler class to reduce excess code
