+++ bundle/src/main/java/com/adobe/acs/commons/rewriter/impl/TemplatedTransformer.java
@@ -1,4 +1,4 @@
-package com.adobe.acs.commons.rewriter;
+package com.adobe.acs.commons.rewriter.impl;
 
 import java.io.IOException;
 import java.io.UnsupportedEncodingException;
@@ -9,6 +9,7 @@ import java.util.Map;
 
 import com.adobe.acs.commons.properties.PropertyAggregatorService;
 import com.adobe.acs.commons.properties.util.TemplateReplacementUtil;
+import com.adobe.acs.commons.rewriter.ContentHandlerBasedTransformer;
 import com.adobe.granite.rest.Constants;
 import com.day.cq.wcm.api.Page;
 import com.day.cq.wcm.api.PageManager;
@@ -16,43 +17,24 @@ import com.day.cq.wcm.api.PageManager;
 import org.apache.sling.api.SlingHttpServletRequest;
 import org.apache.sling.rewriter.ProcessingComponentConfiguration;
 import org.apache.sling.rewriter.ProcessingContext;
-import org.apache.sling.rewriter.Transformer;
-import org.apache.sling.rewriter.TransformerFactory;
-import org.osgi.service.component.annotations.Component;
-import org.osgi.service.component.annotations.Reference;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.xml.sax.Attributes;
-import org.xml.sax.ContentHandler;
-import org.xml.sax.Locator;
 import org.xml.sax.SAXException;
 import org.xml.sax.helpers.AttributesImpl;
 
-@Component(service = TransformerFactory.class, property = {
-        "pipeline.type=templated-transformer"
-    })
-public class TemplatedTransformer implements Transformer, TransformerFactory {
+public class TemplatedTransformer extends ContentHandlerBasedTransformer {
 
     private final Logger log = LoggerFactory.getLogger(this.getClass());
 
-    private ContentHandler contentHandler;
     private Map<String, Object> properties;
-
-    @Reference
-    private PropertyAggregatorService propertyAggregatorService;
-
     private PropertyAggregatorService localService;
 
-    public TemplatedTransformer() {}
+    public TemplatedTransformer() {
+    }
 
     public TemplatedTransformer(PropertyAggregatorService propertyAggregatorService) {
         this.localService = propertyAggregatorService;
-    }
-
-    @Override
-    public Transformer createTransformer() {
-        log.trace("Templated Transformer");
-        return new TemplatedTransformer(propertyAggregatorService);
     }
 
     @Override
@@ -67,7 +49,6 @@ public class TemplatedTransformer implements Transformer, TransformerFactory {
         }
     }
 
-    @Override
     public void startElement(String uri, String localName, String quaName, Attributes atts) throws SAXException {
         if (shouldRun() && localName.equals("a")) {
             AttributesImpl newAttrs = new AttributesImpl(atts);
@@ -88,13 +69,12 @@ public class TemplatedTransformer implements Transformer, TransformerFactory {
                     }
                 }
             }
-            contentHandler.startElement(uri, localName, quaName, newAttrs);
+            getContentHandler().startElement(uri, localName, quaName, newAttrs);
         } else {
-            contentHandler.startElement(uri, localName, quaName, atts);
+            getContentHandler().startElement(uri, localName, quaName, atts);
         }
     }
 
-    @Override
     public void characters(char[] ch, int start, int length) throws SAXException {
         String currentString = new String(ch);
         int placeLength = length;
@@ -120,11 +100,11 @@ public class TemplatedTransformer implements Transformer, TransformerFactory {
             }
         }
 
-        contentHandler.characters(currentString.toCharArray(), start, placeLength);
+        getContentHandler().characters(currentString.toCharArray(), start, placeLength);
     }
 
     private boolean shouldRun() {
-        return localService.isEnabled() && properties != null;
+        return localService != null && properties != null;
     }
 
     private String decode(String input) {
@@ -143,60 +123,5 @@ public class TemplatedTransformer implements Transformer, TransformerFactory {
             log.error("Error encoding object");
         }
         return input;
-    }
-
-    @Override
-    public void endElement(String uri, String localName, String quaName) throws SAXException {
-        contentHandler.endElement(uri, localName, quaName);
-    }
-
-    @Override
-    public void setContentHandler(ContentHandler contentHandler) {
-        this.contentHandler = contentHandler;
-    }
-
-    @Override
-    public void dispose() {
-
-    }
-
-    @Override
-    public void setDocumentLocator(Locator locator) {
-        contentHandler.setDocumentLocator(locator);
-    }
-
-    @Override
-    public void startDocument() throws SAXException {
-        contentHandler.startDocument();
-    }
-
-    @Override
-    public void endDocument() throws SAXException {
-        contentHandler.endDocument();
-    }
-
-    @Override
-    public void startPrefixMapping(String prefix, String uri) throws SAXException {
-        contentHandler.startPrefixMapping(prefix, uri);
-    }
-
-    @Override
-    public void endPrefixMapping(String prefix) throws SAXException {
-        contentHandler.endPrefixMapping(prefix);
-    }
-
-    @Override
-    public void ignorableWhitespace(char[] ch, int start, int length) throws SAXException {
-        contentHandler.ignorableWhitespace(ch, start, length);
-    }
-
-    @Override
-    public void processingInstruction(String target, String data) throws SAXException {
-        contentHandler.processingInstruction(target, data);
-    }
-
-    @Override
-    public void skippedEntity(String name) throws SAXException {
-        contentHandler.skippedEntity(name);
     }
 }
